fix: fix issue with paint tool not working with default color
Fixed a bug where the paint tool was not working with the default color when using the app for the first time. The issue was caused by an error in the JavaScript code. Specifically, the saveColorToLocalStorage function wasn't being called, so the default color wasn't being saved to the color picker.